requirements.txt
Flask==2.3.2
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
bcrypt==4.0.1
qrcode[pil]==7.4.2
uuid==1.30
requests==2.31.0
google-generativeai==0.3.2